Fix minimum distance and add logging

+++ src/algorithms/digi/PulseCombiner.h
@@ -34,7 +34,7 @@ private:
 
   std::vector<std::vector<edm4hep::TimeSeries>> clusterPulses(const std::vector<edm4hep::TimeSeries> pulses) const;
   std::vector<float> sumPulses(const std::vector<edm4hep::TimeSeries> pulses) const;
-  int m_minimum_separation = 1000*dd4hep::ns;
+  float m_minimum_separation = 1000*dd4hep::ns;
   int m_max_time_bins = 10000;
   
 };
